refactor(synthesis): extract distill phase to synthesis/distill.rs
